pyproject.toml
[tool.poetry]
name = "resize-image"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.10.7,<3.11"
fastapi = ">=0.87.0,<0.88.0"
fastapi-camelcase = "^1.0.5"
uvicorn = ">=0.20.0,<0.21.0"
google-cloud-logging = "^3.2.5"
Pillow = "^9.3.0"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
rembg = "^2.0.30"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"
pytest-cov = "^3.0.0"
poethepoet = "^0.11.0"
requests = "^2.27.1"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poe.tasks]
autoformat.sequence = [
    {cmd = "black src tests"},
    {cmd = "isort src tests"},
]

lint.sequence = [
    {cmd = "black --check -v src tests"},
    {cmd = "isort --check -v src tests"},
    {cmd = "flake8 src tests"},
    {cmd = "mypy --pretty src tests"},
]

test.sequence = ["unit", "integration"]

[tool.poe.tasks.unit]
sequence = [
    {cmd = "pytest -vv --cov-report xml --cov-report term --cov=resize_image tests/unit"},
]

[tool.poe.tasks.integration]
sequence = [
    {shell = "pytest -vv --cov-append --cov-report xml --cov-report term --cov=resize_image tests/integration"},
]

[tool.poe.tasks.local-server]
sequence = [
    {shell = "uvicorn resize_image.api:app --reload --port 8011"},
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
